requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
django-heroku # This should not be used
dj-database-url==0.5.0
Django==3.0.3
django-import-export==2.0.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Markdown==3.2.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2 # This should not be used
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
sqlparse==0.3.0
tablib==0.14.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
